requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
flask==3.0.3
python-dotenv==1.0.1

pandas==2.2.2
fake-useragent==1.5.1


